9969: split test out from main test file
