+++ shared/src/business/entities/trialSessions/TrialSession.test.ts
@@ -6,16 +6,6 @@ import {
 } from '../EntityConstants';
 import { TrialSession, isStandaloneRemoteSession } from './TrialSession';
 import { applicationContext } from '../../test/createTestApplicationContext';
-import { isTodayWithinGivenInterval } from '../../utilities/DateHandler';
-
-jest.mock('../../utilities/DateHandler', () => {
-  const originalModule = jest.requireActual('../../utilities/DateHandler');
-  return {
-    __esModule: true,
-    ...originalModule,
-    isTodayWithinGivenInterval: jest.fn(),
-  };
-});
 
 const VALID_TRIAL_SESSION = {
   chambersPhoneNumber: '1234567890',
@@ -616,41 +606,6 @@ describe('TrialSession entity', () => {
       );
 
       expect(trialSession.dismissedAlertForNOTT).toBe(false);
-    });
-  });
-
-  describe('Notice of trial reminder', () => {
-    it('should set isStartDateWithinNOTTReminderRange to false when the trial session is not calendared', () => {
-      const trialSession = new TrialSession(
-        {
-          ...VALID_TRIAL_SESSION,
-          isCalendared: false,
-        },
-        {
-          applicationContext,
-        },
-      );
-
-      expect(trialSession.isStartDateWithinNOTTReminderRange).toBe(false);
-    });
-
-    it('should set isStartDateWithinNOTTReminderRange to true when the trial session is calendared and the trial date falls within the specified date range', () => {
-      const mockIsTodayWithinGivenInterval =
-        isTodayWithinGivenInterval as jest.Mock;
-      mockIsTodayWithinGivenInterval.mockReturnValue(true);
-
-      const trialSession = new TrialSession(
-        {
-          ...VALID_TRIAL_SESSION,
-          isCalendared: true,
-          startDate: '10/10/2020',
-        },
-        {
-          applicationContext,
-        },
-      );
-
-      expect(trialSession.isStartDateWithinNOTTReminderRange).toBe(true);
     });
   });
 });
